refactor: file structure
