+++ src/data_store/mod.rs
@@ -1,1 +1,2 @@
 pub mod store;
+mod value_entry;
